requirements.txt
requests
beautifulsoup4
loguru
python-dotenv